Support for unofficial j2534 linux driver
Will work:
- Tactrix Openport 2.0: https://github.com/dschultzca/j2534
- Machina: https://github.com/rnd-ash/Macchina-J2534
- WQCAN: https://github.com/witoldo7/STM32CAN

Possible to make work:
- Sardine: https://github.com/hackingvolvo/SardineCAN-Win32
